requirements.txt
pandasai
streamlit
matplotlib
chainlit
python-dotenv
openai